+++ Models/ViewModels/IntentViewModel.cs
@@ -9,8 +9,8 @@ namespace FacebookChatbotManagement.Models.ViewModels
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Step { get; set; }
-        public int Exception { get; set; }
+        public double Step { get; set; }
+        public double Exception { get; set; }
         public int DialogId { get; set; }
         public List<PatternViewModel> Patterns { get; set; }
     }
